+++ packages/rma-warranty/src/app/warranty-ui/shared-warranty-modules/service-invoices/add-service-invoice/service-invoice-interface.ts
@@ -33,6 +33,7 @@ export class ServiceInvoiceDetails {
   pos_profile?: string;
   payments?: Payments[];
   set_posting_time?: number;
+  submit?: string;
 }
 
 export class Payments {