+++ packages/rma-server/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.3](https://gitlab.com/castlecraft/excel-rma/compare/rma-server@1.15.2...rma-server@1.15.3) (2021-12-07)
+
+
+### Bug Fixes
+
+* pull warranty claim analysis report ([9659405](https://gitlab.com/castlecraft/excel-rma/commit/9659405ad81b4e8c7cc8329f167069a0b7e199e1))
+
+
+
+
+
 ## [1.15.2](https://gitlab.com/castlecraft/excel-rma/compare/rma-server@1.15.1...rma-server@1.15.2) (2021-10-30)
 
 **Note:** Version bump only for package rma-server